Fix: kernel, english comments

+++ pi-montecarlo/pi_kernel.cl
@@ -2,14 +2,13 @@ __kernel void compute_pi(__global int *count, const int num_samples, __global un
     int id = get_global_id(0);
     int local_count = 0;
     float x, y;
-    
-    // Generatore di numeri casuali semplice
-    unsigned int seed = seeds[id];  // seed diverso per ogni work-item
+
+    unsigned int seed = seeds[id];  // Different seed for each work-item
 
     for (int i = 0; i < num_samples; ++i) {
-        x = (float)rand_r(&seed) / RAND_MAX;
-        y = (float)rand_r(&seed) / RAND_MAX;
-        if (x * x + y * y <= 1.0f) {
+        x = (float)rand_r(&seed) / (float)RAND_MAX;
+        y = (float)rand_r(&seed) / (float)RAND_MAX;
+        if ((x * x + y * y) <= 1.0f) {
             local_count++;
         }
     }
